requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.3
pydantic-settings==2.1.0
